fix ssr on vercel, add check for window, local storage

+++ src/app/cabinet/services/recently-viewed.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable, BehaviorSubject } from 'rxjs';
+import { Observable, BehaviorSubject, throwError } from 'rxjs';
 import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { ProductInterface } from 'src/app/core/services/api-response-types';
 
@@ -18,41 +18,55 @@ export class RecentlyViewedService {
     }
 
     addToRecentlyViewedProds(productId: string):Observable<{message: string}> {
-        const apiUrl = `${this.backendUrl}/add-to-recentlyViewedProds/${productId}` 
-        const token = localStorage.getItem('authToken');
-        const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
-        const options = {
-            headers,
-            withCredentials: true
-        };
-        return this.http.post<{message: string, updatedWishlist: string[]}>(apiUrl, null, options);
+        if (typeof window !== 'undefined' && localStorage) {
+            const apiUrl = `${this.backendUrl}/add-to-recentlyViewedProds/${productId}` 
+            const token = localStorage.getItem('authToken');
+            const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
+            const options = {
+                headers,
+                withCredentials: true
+            };
+            return this.http.post<{message: string, updatedWishlist: string[]}>(apiUrl, null, options);
+        } else {
+            return throwError('Local storage is not available.');
+        }
     }
 
     removeAllRecentlyViewedProds() {
-        const apiUrl = `${this.backendUrl}/remove-recentlyViewedProds` 
-        const token = localStorage.getItem('authToken');
-        const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
-        const options = {
-            headers,
-            withCredentials: true
-        };
-        return this.http.delete<{message: string, updatedProds: []}>(apiUrl, options)
+        if (typeof window !== 'undefined' && localStorage) {
+            const apiUrl = `${this.backendUrl}/remove-recentlyViewedProds` 
+            const token = localStorage.getItem('authToken');
+            const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
+            const options = {
+                headers,
+                withCredentials: true
+            };
+            return this.http.delete<{message: string, updatedProds: []}>(apiUrl, options)
+        } else {
+            return throwError('Local storage is not available.');
+        }
     }
 
     removeOneRecentlyViewedProd(prodId: string) {
-        const apiUrl = `${this.backendUrl}/remove-one-recentlyViewedProd/${prodId}`;
-        const token = localStorage.getItem('authToken');
-        const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
-    
-        return this.http.delete<{message: string}>(apiUrl, { headers, withCredentials: true });
+        if (typeof window !== 'undefined' && localStorage) {
+            const apiUrl = `${this.backendUrl}/remove-one-recentlyViewedProd/${prodId}`;
+            const token = localStorage.getItem('authToken');
+            const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
+            return this.http.delete<{message: string}>(apiUrl, { headers, withCredentials: true });
+        } else {
+            return throwError('Local storage is not available.');
+        }
     }
 
     getRecentlyViewedProds() {
-        const apiUrl = `${this.backendUrl}/recentlyViewedProds`;
-        const token = localStorage.getItem('authToken');
-        const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
-    
-        return this.http.get<{products: any[]}>(apiUrl, { headers, withCredentials: true });
+        if (typeof window !== 'undefined' && localStorage) {
+            const apiUrl = `${this.backendUrl}/recentlyViewedProds`;
+            const token = localStorage.getItem('authToken');
+            const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
+            return this.http.get<{products: any[]}>(apiUrl, { headers, withCredentials: true });
+        } else {
+            return throwError('Local storage is not available.');
+        }
     }
     
 }
