+++ src/app/cabinet/services/wishlist.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable, BehaviorSubject, map } from 'rxjs';
+import { Observable, BehaviorSubject, map, throwError } from 'rxjs';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { ProductInterface } from 'src/app/core/services/api-response-types';
 import { ISearchStatus } from 'src/app/core/services/api-response-types';
@@ -23,56 +23,61 @@ export class WishlistService {
 
     private wishlistItemsSubject = new BehaviorSubject<any[]>([]);
     wishlistItems$: Observable<any[]> = this.wishlistItemsSubject.asObservable();
-
     total: number = 0
-
     sortOptions: string[] = ['За датою додавання', 'Від дешевих до дорогих', 'Від дорогих до дешевих']
     activeSortOption: string = this.sortOptions[0]
 
     setWishlistItems(data: ProductInterface[]) {
-        console.log('setWishlistItems starts')
         this.wishlistItemsSubject.next(data)
         this.sortWishlist(this.activeSortOption)
         this.calculateTotalPrice(data) 
     }
 
     addToWishlist(productId: string):Observable<{message: string, updatedWishlist: ProductInterface[]}> {
-        const apiUrl = `${this.backendUrl}/add-to-wishlist/${productId}` 
-        const token = localStorage.getItem('authToken');
-        const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
-        const options = {
-            headers,
-            withCredentials: true
-        };
-        return this.http.post<{message: string, updatedWishlist: ProductInterface[]}>(apiUrl, null, options).pipe(
-            map((response: any) => {
-                if (response && response.updatedWishlist) {
-                    response.updatedWishlist = response.updatedWishlist.map((wishlistItem: WishlistItemResponse) => ({
-                        ...wishlistItem.product, 
-                        addedDate: wishlistItem.addedDate 
-                    }));
-                }
-                return response;
-            })
-        )
+        if (typeof window !== 'undefined' && localStorage) {
+            const apiUrl = `${this.backendUrl}/add-to-wishlist/${productId}` 
+            const token = localStorage.getItem('authToken');
+            const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
+            const options = {
+                headers,
+                withCredentials: true
+            };
+            return this.http.post<{message: string, updatedWishlist: ProductInterface[]}>(apiUrl, null, options).pipe(
+                map((response: any) => {
+                    if (response && response.updatedWishlist) {
+                        response.updatedWishlist = response.updatedWishlist.map((wishlistItem: WishlistItemResponse) => ({
+                            ...wishlistItem.product, 
+                            addedDate: wishlistItem.addedDate 
+                        }));
+                    }
+                    return response;
+                })
+            )
+        } else {
+            return throwError('Local storage is not available.');
+        }
     }
 
     removeFromWishlist(productIds: string[]): Observable<WishlistResponse> {
-        const apiUrl = `${this.backendUrl}/remove-from-wishlist`;
-        const token = localStorage.getItem('authToken');
-        const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
-        const data = { productIds };
-        return this.http.delete<WishlistResponse>(apiUrl, { headers, body: data, withCredentials: true }).pipe(
-            map((response: any) => {
-                if (response && response.updatedWishlist) {
-                    response.updatedWishlist = response.updatedWishlist.map((wishlistItem: WishlistItemResponse) => ({
-                        ...wishlistItem.product, 
-                        addedDate: wishlistItem.addedDate 
-                    }));
-                }
-                return response;
-            })
-        )
+        if (typeof window !== 'undefined' && localStorage) {
+            const apiUrl = `${this.backendUrl}/remove-from-wishlist`;
+            const token = localStorage.getItem('authToken');
+            const headers = new HttpHeaders({ Authorization: `Bearer ${token}` });
+            const data = { productIds };
+            return this.http.delete<WishlistResponse>(apiUrl, { headers, body: data, withCredentials: true }).pipe(
+                map((response: any) => {
+                    if (response && response.updatedWishlist) {
+                        response.updatedWishlist = response.updatedWishlist.map((wishlistItem: WishlistItemResponse) => ({
+                            ...wishlistItem.product, 
+                            addedDate: wishlistItem.addedDate 
+                        }));
+                    }
+                    return response;
+                })
+            )
+        } else {
+            return throwError('Local storage is not available.');
+        }
     }
 
     sortWishlist(sortType: string) {
